% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climograph.R
\name{climograph}
\alias{climograph}
\title{Create Climograph}
\usage{
climograph(
  name = "worldclim",
  res = 10,
  lon = 39.5,
  lat = 40.5,
  path = "",
  download = TRUE,
  style = "normal",
  save = FALSE,
  filename = "Rplots.pdf",
  height = 7,
  width = 7,
  bg = "transparent"
)
}
\arguments{
\item{name}{character. Name of dataset, default is worldclim.}

\item{res}{numeric. One of 0.5, 1, 5, 10. Resolution of worldclim data to use, default is 10.}

\item{lon}{Longitude of location you want to create climograph for.}

\item{lat}{Latitude of location you want to create climograph for.}

\item{path}{Character. Path name indicating where to store the data. Default is the current working directory.}

\item{download}{Logical. If TRUE data will be downloaded if not locally available}

\item{style}{character. normal, plotly or ggplot. Defaults to normal.}

\item{save}{True/False. If Plot should be saved to file.}

\item{filename}{the name of the output file.}

\item{height}{the height of the device in inches.}

\item{width}{the width of the device in inches.}

\item{bg}{the initial background colour, default is transparent.}
}
\value{
Returns a climograph plot  with coordinates (x,y) and corresponding raster values.
}
\description{
Function to plot climograph of a given location in the world.
}
\examples{
\dontrun{
climograph(lon=39.5, lat=40.5)
}
}
