% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap2-package.R
\docType{data}
\name{continents}
\alias{continents}
\title{Outline of the world's continents}
\format{
A sf-data.frame with 8 observations.
}
\usage{
data(continents)
}
\description{
Outline of the world split into the different continents
}
\details{
GADM outline of the world split into the different continents
}
