% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_transform.R
\name{df_transform}
\alias{df_transform}
\title{Transform projection of data.frame}
\usage{
df_transform(data, coords, crs.in, crs.out)
}
\arguments{
\item{data}{data.frame object}

\item{coords}{vector of the two coordinate column names}

\item{crs.in}{original projection of data.frame, standard is WGS1984.}

\item{crs.out}{output projection of data.frame, which is returned.}
}
\value{
A data.frame
}
\description{
Change the projection of x and y values (spatial coordinates) 
             of a data.frame, make sure your data only includes x,y and z 
             and no additional categorical variables
}
\details{
df_transform internally converts the data.frame into a spatial object, 
than transforms the projection of the data and 
converts the final spatial object back into a data.frme.
}
\examples{
data(meuse, package="sp")
meuse <- df_transform(meuse, 
                      crs.in=sp::CRS("+init=epsg:28992"), 
                      crs.out=sp::CRS("+init=epsg:4326"))

}
