% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceToCoast.R
\name{distanceToCoast}
\alias{distanceToCoast}
\title{Calculate distance to coast}
\usage{
distanceToCoast(ncol, nrow, extent='', inverse=FALSE, ...)
}
\arguments{
\item{ncol}{Number of columns used for creating distanceToCoast raster}

\item{nrow}{Number of rows used for creating distanceToCoast raster}

\item{extent}{Extent used for creating distanceToCoast raster}

\item{inverse}{Logical. If TRUE the distance to land will be calculated}

\item{...}{Additional required (!) parameters. 
These are data set specific. See Details}
}
\value{
A Raster* layer
}
\description{
Create a raster layer with the distance to coast for the sea or 
land (if inverse=TRUE).
}
\details{
distanceToCoast internally loads the outline data, 
which is also included in this package. It then creates a raster of the outline shapefile 
using the rasterize function according to the specified resolution and extent provided 
and then calculates the distance using the distance function from the raster package.
}
\examples{
distSea <- distanceToCoast(ncol=36, nrow=18)
distLand <- distanceToCoast(ncol=36, nrow=18, inverse=TRUE)

}
\references{
\url{http://www.gadm.org}
}
\keyword{spatial}
