% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNDVI.R
\name{downloadNDVI}
\alias{downloadNDVI}
\title{Download ndvi3g data}
\usage{
downloadNDVI(startyear = 1981, endyear = 2013, version = "v0", path = getwd())
}
\arguments{
\item{startyear}{A year between 1981 and 2015.}

\item{endyear}{A year between 1981 and 2015, should be larger than startyear.}

\item{version}{One of "v0" or "v1", otherwise function automatically uses "v1".}

\item{path}{Path where data is downloaded to. If not available a subfolder with the version name will be automatically created.}
}
\value{
The function just downloads the files to the specified path.
}
\description{
Function to download AVHRR GIMMS ndvi3g data for a specific timeperiod of your choice.
Data is available globally and for the timeperiod 1981 - 2015.
}
\examples{
\dontrun{
downloadNDVI(startyear=1981, endyear=2013, version="v0")
downloadNDVI(startyear=1981, endyear=2015, version="v1")
}
}
