% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData2.R
\name{getData2}
\alias{getData2}
\title{Get geographic data}
\usage{
getData2(name, download=TRUE, extent='', path='', ...)
ccodes()
}
\arguments{
\item{name}{Data set name, currently supported are 'GADM', 'countries', 'SRTM', 'alt', 'CMIP5', 'WDPA' and 'worldclim'. 
See Details for more info. GADM', 'countries', 'SRTM', 'alt', 'CMIP5' and 'worldclim' are further supported by the getData() function of the raster package.}

\item{download}{Logical. If TRUE data will be downloaded if not locally available}

\item{extent}{extent object. Spatial object extent delimiting the area you are interested in.}

\item{path}{Character. Path name indicating where to store the data. Default is the current working directory}

\item{...}{Additional required (!) parameters. These are data set specific. See Details}
}
\value{
A spatial object (Raster* or sf`*)
}
\description{
Get geographic data for anywhere in the world. Data are read from files that are first downloaded if necessary. 
Function ccodes returns country names and the ISO codes

This function extends the getData function from the raster package to download and
access various other geographic data (BioOracle, GEBCO, MarSpec, OceanColor, WOA, WDPA, gimms3g).
}
\details{
'alt' stands for altitude (elevation); the data were aggregated from SRTM 90 m resolution data between -60 and 60 latitude. 'GADM' is a database of global administrative boundaries. 'worldclim' is a database of global interpolated climate data. 'SRTM' refers to the hole-filled CGIAR-SRTM (90 m resolution). 

'countries' has polygons for all countries at a higher resolution than the 'wrld_simpl' data in the maptools package . 

If \code{name} is 'alt' or 'GADM' you must provide a 'country=' argument. Countries are specified by their 3 letter ISO codes. Use getData('ISO3') to see these codes. In the case of GADM you must also provide the level of administrative subdivision (0=country, 1=first level subdivision). In the case of alt you can set 'mask' to FALSE. If it is TRUE values for neighbouring countries are set to NA. For example:
\code{getData('GADM', country='FRA', level=1)}
\code{getData('alt', country='FRA', mask=TRUE)}
If \code{name} is 'SRTM' you must provide 'lon' and 'lat' arguments (longitude and latitude). These should be single numbers somewhere within the SRTM tile that you want.
\code{getData('SRTM', lon=5, lat=45)}
If \code{name='worldclim'} you must also provide arguments \code{var}, and a resolution \code{res}. Valid variables names are 'tmin', 'tmax', 'prec' and 'bio'. Valid resolutions are 0.5, 2.5, 5, 10, 30 and 60 (minutes of a degree). In the case of \code{res=0.5}, you must also provide a \code{lon} and \code{lat} argument for a tile or an extent argument; for the lower resolutions global data will be downloaded. In all cases there are 12 (monthly) files for each variable except for 'bio' which contains 19 files.
\code{getData('worldclim', var='tmin', res=0.5, lon=5, lat=45)}
\code{getData('worldclim', var='bio', res=10)}
To get (projected) future climate data (CMIP5), you must provide arguments \code{var} and \code{res} as above. Only resolutions 2.5, 5, and 10 are currently available. 
In addition, you need to provide \code{model}, \code{rcp} and \code{year}. For example,
\code{getData('CMIP5', var='tmin', res=10, rcp=85, model='AC', year=70)}
function (var, model, rcp, year, res, lon, lat, path, download = TRUE) 
'model' should be one of "AC", "BC", "CC", "CE", "CN", "GF", "GD", "GS", "HD", "HG", "HE", "IN", "IP", "MI", "MR", "MC", "MP", "MG", or  "NO".
'rcp' should be one of 26, 45, 60, or 85.
'year' should be 50 or 70
Not all combinations are available. See www.worldclim.org for details.

If \code{name='WDPA'} you can provide a 'country=' argument to download the current protected areas of the specific country.
\code{getData('WDPA', country='LIE')}
}
\examples{

\dontrun{
getData2(name = "worldclim", res=30, var="tmin")
getData2(name="worldclim", res=10, var="tmin", extent=c(8, 14, 47, 51))
getData2(name="worldclim", res=0.5, var="tmin", extent=c(0, 20, 20, 50))
getData2(name="CMIP5", var="tmin", res=30, rcp=85, model="AC", year=70)
getData2(name="CMIP5", var="tmin", res=10, rcp=85, model="AC", year=70, extent=c(0, 20, 20, 50))
getData2(name = "WDPA", country="LIE")
}

}
\references{
\url{http://www.worldclim.org}
\url{http://www.gadm.org}
\url{http://srtm.csi.cgiar.org/}
\url{http://diva-gis.org/gdata}
\url{http://protectedplanet.net}
}
\keyword{spatial}
