% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap2-package.R
\docType{package}
\name{ggmap2 package}
\alias{ggmap2 package}
\alias{ggmap2package}
\title{ggmap2 - R package to easily create nice maps}
\description{
R package to easily create nice maps
}
\author{
RS-eco
}
\keyword{package}
