% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap2.R
\name{ggmap2}
\alias{ggmap2}
\title{Create map(s) of one raster layer or a raster stack}
\usage{
ggmap2(
  data,
  name = "",
  split = FALSE,
  trans = "identity",
  facet_grid = FALSE,
  crop = TRUE,
  grid = FALSE,
  colours = colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan", "#7FFF7F", "yellow",
    "#FF7F00", "red", "#7F0000"))(255),
  varnames = NA,
  subnames = "",
  outline = NA,
  extent = NA,
  ncol = 2,
  long = FALSE,
  country = TRUE,
  save = FALSE,
  filename = "Rplots.pdf"
)
}
\arguments{
\item{data}{raster layer, raster stack or dataframe}

\item{name}{character, name of legend}

\item{split}{logical, should multiple plots have individual legends or not}

\item{trans}{transform data}

\item{facet_grid}{logical, default to FALSE, if TRUE facet_grid is used for plotting}

\item{crop}{logical, default to TRUE, not implemented right now}

\item{grid}{logical, default to FALSE, if TRUE a grid is added to the plot}

\item{colours}{specify colourPalette used for ggplot2::scale_fill_gradientn()}

\item{varnames}{character vector with names of data}

\item{subnames}{character vector with subnames of data}

\item{outline}{default is NA}

\item{extent}{defaults to NA}

\item{ncol}{specify number of columns for plotting with facet_wrap or ..., 
not needed if facet_grid is TRUE.}

\item{long}{defaults to FALSE. Specify if data is in long format or not.}

\item{country}{defaults to TRUE. Add country outline to plot.}

\item{save}{defaults to FALSE. Specify if plot should be written to file.}

\item{filename}{If save=TRUE you can specify the filename of your plot here}
}
\value{
ggplot2 object
}
\description{
Read a global raster layer or raster stack and plot on a map
}
\examples{
# Load raster library
library(raster)
# Create data
r <- raster(nrow=90, ncol=180)
r <- setValues(r, rnorm(ncell(r), 0, 1))
#Create plots
ggmap2(r, name="Single", split=FALSE, ncol=1)
ggmap2(r, name="Single", split=FALSE, ncol=1, extent=c(-20, 70, -50, 80))
}
