% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRegression.R
\name{ggRegression}
\alias{ggRegression}
\title{Get GIMMS3g NDVI time series data}
\usage{
ggRegression(
  data,
  vars = c("slope", "R2", "adjR2", "pValue"),
  save = FALSE,
  filename = NULL,
  colour = colorRampPalette(c("#00007F", "blue", "#007FFF", "cyan", "#7FFF7F", "yellow",
    "#FF7F00", "red", "#7F0000"))(255)
)
}
\arguments{
\item{data}{rasterStack with date information}

\item{vars}{currently only takes all variables}

\item{save}{specify if you want to save the plot to file or not, default = FALSE}

\item{filename}{define filename if plot should be saved to file}

\item{colour}{specify colour palette of ggplot2::scale_fill_gradientn()}
}
\value{
ggplot2 object
}
\description{
Get GIMMS3g NDVI time series data
}
\examples{
\dontrun{
library(ggmap2)
dat_v0 <- ndvi3g(extent = c(10, 30, 30, 50), version="v0", startyear=1981, 
                 endyear=1985, path="/media/matt/Data/Documents/Wissenschaft/Data/GIMMS")
ggRegression(data=dat_v0)
}
}
