% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap2-package.R
\docType{data}
\name{largeislands}
\alias{largeislands}
\title{Outline of the world's islands with a size larger than 5 km2}
\format{
A sf-data.frame with 8365 observations and 20 variables.
}
\usage{
data(largeislands)
}
\description{
GADM of the world's islands with a size larger than 5 km2
}
\details{
This shapefile contains the outline of the world's islands.
}
