% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi3g.R
\name{ndvi3g}
\alias{ndvi3g}
\title{Load ndvi3g data}
\usage{
ndvi3g(
  path,
  extent = c(-180, 180, -90, 90),
  startyear = 1981,
  endyear = 2015,
  snap = "near",
  version = "v1",
  output = "ndvi"
)
}
\arguments{
\item{path}{Path to file directory where data is stored, should be same as used in downloadNDVI().}

\item{extent}{An extent object.}

\item{startyear}{A year between 1981 and 2015.}

\item{endyear}{A year between 1981 and 2015, should be larger than startyear.}

\item{snap}{see raster::crop}

\item{version}{One of "v0" or "v1", otherwise function automatically uses "v1".}

\item{output}{specify the output variable, default is "ndvi". v0 provides two options: flag and ndvi. v1 provides two options: ndvi and percentile.}
}
\value{
raster stack with the extent of \code{extent} and layers for each year from \code{startyear} till \code{endyear}.
}
\description{
Function to dextract AVHRR GIMMS ndvi3g data for a specific 
timeperiod and area of your choice. To download the data see the downloadNDVI() function.
Data is available globally and for the timeperiod 1981 - 2015.
}
\examples{
\dontrun{
dat_v0 <- ndvi3g(extent = c(10, 30, 30, 50), version="v0", 
                 startyear=1981, endyear=1982)
flag_dat_v0 <- ndvi3g(extent = c(10, 30, 30, 50), version="v0", startyear=1981, 
                      endyear=1982, output="flag")
dat_v1 <- ndvi3g(extent = c(10, 30, 30, 50), version="v1", startyear=1981, endyear=1982)
perc_dat_v1 <- ndvi3g(extent = c(10, 30, 30, 50), version="v1", startyear=1981, 
                      endyear=1982, output="percentile")
}
}
