% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmap2-package.R
\docType{data}
\name{outline}
\alias{outline}
\title{Outline of the world}
\format{
\code{sfc_Multipolygon}.
}
\usage{
data(outline)
}
\description{
Shapefile of the world's land outline
}
\details{
This shapefile contains information about the world's land outline
}
