% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_map.R
\name{theme_map}
\alias{theme_map}
\title{Clean theme for maps}
\usage{
theme_map(base_size = 9, base_family = "")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
A clean theme that is good for displaying maps.
}
\examples{
library(ggplot2)
ggplot(faithfuld, aes(waiting, eruptions)) + geom_raster(aes(fill = density)) + 
       theme_map()
}
